requirements.txt
langchain==0.3.25
langchain-openai==0.3.17
langgraph==0.4.5
fastapi==0.115.12
uvicorn==0.34.2
python-dotenv==1.1.0
pydantic==2.11.4
